DropBox.m: eliminate compiler warning
NSBorderType is of type NSUInteger which is either 32
or 64 bits in size.  Replace the '%d' formatter with
'%u' and cast to unsigned to avoid the warning.

+++ source/DropBox.m
@@ -95,7 +95,7 @@
 
     if (borderType < 0 || borderType > 3)
     {
-        fprintf(stderr, "DropBox: invalid NSBorderType: %d\n", borderType);
+        fprintf(stderr, "DropBox: invalid NSBorderType: %u\n", (unsigned)borderType);
         return;
     }
